Rename XmlEntity class to XmlUtils
Also moved it to Utils namespace.
